Move tester runtime classes to own package
